+++ Assets/UniVRM-1.0/Components/Editor/BlendShape/ReorderableBlendShapeBindList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0bfe7ed77bf5e3b46ba4fe300132ce10
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
