+++ Assets/UniVRM-1.0/Components/Editor/BlendShape/ReorderableMaterialBindList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6f6fcbf32b1c674aa717d3a7daabf23
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
